+++ backend/submodule/rs_device_resources.py
@@ -1,0 +1,133 @@
+from enum import Enum
+
+class ModuleType(Enum):
+    CLOCKING = 0
+    FABRIC_LE = 1
+    DSP = 2
+    BRAM = 3
+    IO = 4
+    PERIPHERAL_PROCESSING = 5
+    REGULATOR = 6
+
+class RsDeviceResources:
+
+    def __init__(self, device):
+        self.device = device
+        self.modules = [None, None, None, None, None, None, None]
+
+    def get_attr(self, name) -> int:
+        return int(self.device.resources[name].num)
+
+    def get_num_Clocks(self):
+        # NOTE:
+        #   The no. of clock resource is not available in device.xml.
+        #   Thus, hardcode here till this is available 
+        return 16
+
+    def get_num_PLLs(self) -> int:
+        # NOTE:
+        #   The no. of PLL resource is not available in device.xml.
+        #   Thus, hardcode here till this is available 
+        series = self.device.series
+        if series in ('Gemini', 'Orion', 'Lyra', 'Vega'):
+            return 4
+        elif series == 'Virgo':
+            return 2
+
+    def get_num_DSP_BLOCKs(self) -> int:
+        # return self.get_attr('dsp')
+        return 200 # overwrite for test purpose
+
+    def get_num_18K_BRAM(self) -> int:
+        # return self.get_attr('bram')
+        return 256 # overwrite for test purpose
+
+    def get_num_36K_BRAM(self) -> int:
+        # return self.get_attr('bram')
+        return 128 # overwrite for test purpose
+
+    def get_series(self):
+        return self.device.series
+
+    def get_device_name(self):
+        return self.device.name
+
+    def get_package(self):
+        return self.device.package
+
+    def get_speedgrade(self):
+        return self.device.speedgrade
+
+    def get_logic_density(self):
+        return ''
+
+    def get_temperature_grade(self):
+        return "Industrial (-40 to 100 °C)"
+
+    def get_num_LUTs(self) -> int:
+        return self.get_attr('lut')
+
+    def get_num_FFs(self) -> int:
+        return self.get_attr('ff')
+
+    def get_CLK_CAP(self) -> float:
+        # todo: read from power data. Coeffient to calculate clock block power
+        return 0.00001
+
+    def get_CLK_INT_CAP(self) -> float:
+        # todo: read from power data. Coeffient to calculate clock interconnect power
+        return 0.00000003
+
+    def get_PLL_INT(self) -> float:
+        # todo: read from power data. Coeffient to calculate PLL power (VCC Core)
+        return 0.0009
+
+    def get_PLL_AUX(self) -> float:
+        # todo: read from power data. Coeffient to calculate PLL power (Aux Int)
+        return 0.01
+
+    def get_LUT_CAP(self) -> float:
+        # todo: read from power data. Coeffient to calculate FLE block power
+        return 0.0000003
+
+    def get_FF_CAP(self) -> float:
+        # todo: read from power data. Coeffient to calculate FLE block power
+        return 0.00000035
+
+    def get_FF_CLK_CAP(self) -> float:
+        # todo: read from power data. Coeffient to calculate FLE block power
+        return 2.91375291375291E-09
+
+    def get_LUT_INT_CAP(self) -> float:
+        # todo: read from power data. Coeffient to calculate FLE interconnect power
+        return 0.00000002
+
+    def get_FF_INT_CAP(self) -> float:
+        # todo: read from power data. Coeffient to calculate FLE interconnect power
+        return 0.00000004
+
+    def get_VCC_CORE(self) -> float:
+        # todo: should read from regulator module
+        return 0.8
+
+    def get_VCC_AUX(self) -> float:
+        # todo: should read from regulator module
+        return 1.8
+
+    def register_module(self, modtype, module):
+        self.modules[modtype.value] = module
+        return module
+
+    def get_modules(self):
+        return self.modules
+
+    def get_module(self, modtype):
+        return self.modules[modtype.value]
+
+    def get_clock(self, clkname):
+        clock_module = self.get_module(ModuleType.CLOCKING)
+        if clock_module != None:
+            for clock in clock_module.get_all():
+                if clock.port == clkname:
+                    return clock
+        return None
